rename batch test to stress test
